+++ analysis/src/main/java/cz/vutbr/fit/group/verifit/oslc/analysis/outputParser/ParserManager.java
@@ -10,8 +10,10 @@
 
 package cz.vutbr.fit.group.verifit.oslc.analysis.outputParser;
 
+import java.lang.reflect.Constructor;
 import java.net.URI;
 import java.net.URISyntaxException;
+import java.nio.file.Paths;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.LinkedList;
@@ -21,6 +23,13 @@ import java.util.Set;
 
 import org.eclipse.lyo.oslc.domains.auto.Contribution;
 import org.eclipse.lyo.oslc4j.core.model.Link;
+
+import cz.vutbr.fit.group.verifit.jsem.ExtensionInfoQuery;
+import cz.vutbr.fit.group.verifit.jsem.ExtensionManager;
+import cz.vutbr.fit.group.verifit.jsem.IExtension;
+import cz.vutbr.fit.group.verifit.jsem.IExtensionWithInfo;
+import cz.vutbr.fit.group.verifit.oslc.analysis.automationPlans.AutomationPlanConfManager;
+import cz.vutbr.fit.group.verifit.oslc.analysis.properties.VeriFitAnalysisProperties;
 
 
 public final class ParserManager {
@@ -44,6 +53,46 @@ public final class ParserManager {
     
     private void loadParsers() {
     	//this.toolParsers.put("infer", new NoFileValuesParser()); // TODO tmp instead of actual plugins
+    	
+    	// initialize the ExtensionManager
+    	ExtensionManager extMgr = new ExtensionManager();
+    	extMgr.registerBuiltinExtensions();
+    	extMgr.refreshExtensionInfoLoaders();
+    	extMgr.loadExtensionInfo(Paths.get(VeriFitAnalysisProperties.PLUGIN_PARSER_CONF_PATH));
+    	
+    	// load parsers for every Automation Plan
+    	AutomationPlanConfManager autoPlanConfManager = AutomationPlanConfManager.getInstance();
+    	Set<String> autoPlans = autoPlanConfManager.getAllAutoPlanIds();
+    	for (String id : autoPlans)
+    	{
+    		if (id.equals("dummy"))
+    		{
+    			// skip the dummy tool (that should always use the default parser)
+    		}
+    		else
+    		{
+    			// find all parser extensions for the given Automation Plan
+    			ExtensionInfoQuery query = new ExtensionInfoQuery();
+    			query.contains("implements", IParser.class.getName());
+    			query.contains("tool", id);
+    			List< IExtension > exts = extMgr.findExtension(query);
+    			
+    			// select a plugin parser for the given Automation Plan TODO move into input parameters (dynamic selection)
+    			if (exts.size() == 1)
+    			{
+    				IExtension ext = exts.get(0);
+    				IParser pluginParser = (IParser) ext;
+    				this.toolParsers.put(id, pluginParser);
+    			}
+    			else
+    			{
+    				// TODO future features
+    				System.out.println("Warning: Failed to load pluginParser for Automation Plan: \"" + id + "\"\n"
+    						+ "   Either no parser was found, or there was more than one.\n"
+    						+ "   Using the default parser.");
+    			}
+    		}
+    	}
     }
     
     /**